requirements.txt
hydra-core==1.3
numpy==2.2.3
matplotlib==3.10.0
pandas==2.2.3